bird+ jump (click but not Keypress )
WHYWHYWHY

+++ src/model/Background.java
@@ -14,8 +14,8 @@ import javafx.scene.image.Image;
  * @author dennisli
  */
 public class Background extends GraphicalObjs{
-    private final int BG_INIT_X = 0;
-    private final int BG_INIT_Y = 0;
+    private final double BG_INIT_X = 0;
+    private final double BG_INIT_Y = 0;
     private final double BG_VELOCITY_X = 0.005;
     private final double BG_VELOCITY_Y = 0.0;
